+++ src/Page.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Buttons from './Buttons';
+import Counter from './Counter';
+
+function Page({ count, onClick }) {
+  return (
+    <div>
+      <p>Hello,world!</p>
+      <Counter
+        count={count}
+        onClick={onClick}
+      />
+      <Buttons />
+    </div>
+  );
+}
+
+export default Page;
